% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splott.R
\name{colorScaleInternals}
\alias{colorScaleInternals}
\alias{combineTwoScales}
\alias{makeDivScale}
\alias{twoColorScale}
\alias{ThreeColorDivScale}
\title{Internal functions for building color scales}
\usage{
combineTwoScales(n, scale1, scale2, args1, args2)

makeDivScale(n, scale, ...)

twoColorScale(n, col1 = grDevices::rgb(0.8, 0.8, 0.8), col2 = "green")

ThreeColorDivScale(
  n,
  vals = NULL,
  valRange = NULL,
  center = NULL,
  col1 = "green",
  col2 = grDevices::rgb(0.8, 0.8, 0.8),
  col3 = "blue"
)
}
\arguments{
\item{n}{Number of levels in the color scale}

\item{...}{Other arguments to the color scale}

\item{rev}{Whether or not to reverse the color scale}

\item{ggplot}{Whether to output a ggplot2-based color scale or based R color
vector}
}
\description{
Internal functions for building color scales
}
\details{
Internal functions for building color scales
}
\examples{
# TODO
}
\seealso{
\code{\link{nonlinearScales}}
}
\author{
John Paige
}
