% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splott.R
\name{colorScaleUtilities}
\alias{colorScaleUtilities}
\alias{getColorsFromScale}
\alias{centerColorScale}
\title{Gets colors from color scale}
\usage{
getColorsFromScale(
  vals,
  valRange = NULL,
  center = NULL,
  cols,
  scaleFun = function(x) {
     x
 },
  forceValuesInRange = FALSE
)

centerColorScale(
  n,
  vals = NULL,
  valRange = NULL,
  center,
  colScale,
  scaleFun = function(x) {
     x
 },
  compressed = FALSE,
  ...
)
}
\arguments{
\item{vals}{Response. Determines colors from the scale}

\item{valRange}{Range of responses for setting the scale}

\item{center}{Center of the range of responses set setting diverging color
scales}

\item{cols}{A vector of colors forming color scale}

\item{scaleFun}{How to scale the color scale. For example, log}

\item{colScale}{a function taking 'n' as input and returns a color scale. See
\code{?colorScales} for formatting}

\item{compressed}{Rather than cutting off a color scale to make the center at
the desired point, compress the scale on the shorter side so it changes color
faster.}

\item{...}{additional arguments passed to \code{colScale}}

\item{forceColorsInRange}{Whether or not to force colors in the plotted
range. Useful if you have a value outside of the range or that is NA after
being transformed via the scale that you still want to plot at the edge of
the color scale}
}
\description{
Gets colors of responses from a color scale, potentially based on scale
transformations and with centering
}
\details{
Use \code{getColorsFromScale} when using a log, logit, or some other
fancy transformation of a color scale. Given data to be assigned colors from
scale (vals), the range of the values, the center of the value range (only
for diverging scales), the color scale as a vector of colors (cols), the
scale of the color scale (e.g. identity, log, logit), and whether or not to
force the colors into valRange, \code{getColorsFromScale} returns the colors
associated with each value in vals.

\code{centerColorScale} centers a color scale at its midpoint. Returns vector of
the centered color scale. Useful when using diverging scales centered at 0
for data with asymmetric range colScale a function taking 'n' as input and
returns a color scale centered in the middle
}
\examples{

set.seed(123)

# construct data:
testX = exp(stats::rnorm(100))

# construct centered color scale on log scale
test = centerColorScale(64, testX, center=1, colScale=redBlueDivCols, scaleFun=log)

# get the colors associated with each value
testCols = getColorsFromScale(testX, center=1, cols=test, scaleFun=log)

# plot the result
plot(testX, col=testCols, pch=19)
}
\seealso{
\code{\link{nonlinearScales}}, \code{\link{nonlinearTicks}},
\code{\link{colorScales}}
}
\author{
John Paige
}
