% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splott.R
\name{nonLinearTicks}
\alias{nonLinearTicks}
\alias{getLogitScaleTicks}
\title{Generates pretty tick marks for nonlinear scales}
\usage{
getLogitScaleTicks(x, nint = 3)
}
\arguments{
\item{x}{Data to base ticks from prior to scaling}

\item{nint}{Approximate desired number of ticks in each part of the scale.
See \code{n} in ?pretty\code{and}nint\verb{in}grDevices::axisTicks`}

\item{nintSmall}{Approximate desired number of ticks in the small
magnitude/linear part of the scale}

\item{nintLarge}{Approximate desired number of ticks in the large
magnitude/log part of the scale}

\item{smallThresh}{Threshold in magnitude dividing the small/linear and
large/log portions of the scale}
}
\description{
Similar to \code{pretty}, but for nonlinear scale tick marks
}
\details{
Use these functions when using nonlinear transformations of a color scale.
Given data to be assigned colors from scale (x), and the number of tick
intervals/integration points (nint), these functions returns a pretty set of
tick marks for a legend.

Each function assumes a different typpe of scale, such as a logit scale for
data in (0,1), a log scale for data in (0, infty), and a log1p-like scale for
data in (-infty, infty) with large outliers in either direction.

For a log1p-based scale (log(x+1)), the scale is approximately linear near 0,
and approximately log otherwise. Hence, the final tick marks are constructed
by first generating a tick marks for a linear scale for small magnitude
values (between \code{-smallThresh} and \code{smallThresh}) with approximately
\code{nintSmall} tick marks, and concatenating this with another set of tick marks
assuming a log scale for magnitudes greater than \code{smallThresh}.
}
\examples{
set.seed(123)
x = matrix(rnorm(100), ncol=2)
z = sign(runif(50)-0.5) * rlnorm(50)

ticks = getLog1pLikeScaleTicks(z)
splot(x, z, colScale=redBlueDivCols, centerScale=TRUE, 
      scaleFun=scaleFunLog1p, ticks=ticks)
}
\seealso{
\code{\link{grDevices::axisTicks}}, \code{\link{pretty}},
\code{\link{nonlinearScales}}
}
\author{
John Paige
}
