% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splott.R
\name{splot}
\alias{splot}
\title{plot points spatially with colors}
\usage{
splot(
  x,
  y,
  z,
  zlim = NULL,
  col = NULL,
  nlevel = 64,
  colScale = blueGreenYellowSeqCols,
  center = 0,
  centerScale = FALSE,
  compressed = FALSE,
  legend.mar = 7,
  new = TRUE,
  scaleFun = function(x) {
     x
 },
  n.ticks = 5,
  min.n = 5,
  ticks = NULL,
  tickLabels = NULL,
  legend.width = 1.2,
  addColorBar = TRUE,
  legendArgs = list(),
  leaveRoomForLegend = TRUE,
  forceColorsInRange = FALSE,
  pch = 19,
  colorName = c("col", "bg"),
  ...
)
}
\arguments{
\item{x}{Horizontal spatial coordinates}

\item{y}{Vertical spatial coordinates}

\item{z}{Response. Determines color of the points}

\item{zlim}{Range of the response}

\item{col}{A vector of colors determining the color scale for plotting (it is
recommended to use the \code{colScale} argument instead)}

\item{nlevel}{Number of colors to include in generated color scale if \code{col}
not specified}

\item{colScale}{A vector of colors determining the color scale for plotting}

\item{center}{Value at which to center the color scale if \code{centerScale} set
to TRUE}

\item{centerScale}{Whether to center the color scale at \code{center} (e.g. for
diverging color scales)}

\item{compressed}{Rather than cutting off a color scale to make the center at
the desired point, compress the scale on the shorter side so it changes color
faster.}

\item{legend.mar}{(as legend.args) see \code{\link{fields::image.plot}}}

\item{new}{Whether to make a new plot or add to the current one}

\item{scaleFun}{How to scale the color scale. For example, log}

\item{n.ticks}{Approximate number of ticks in color scale. See
\code{\link{pretty}}}

\item{min.n}{Minimum number of ticks in color scale. See
\code{\link{pretty}}}

\item{legend.width}{(as legend.args) see \code{\link{fields::image.plot}}}

\item{addColorBar}{Whether to add the color bar/legend}

\item{legendArgs}{Other arguments to legend.args. See}

\item{leaveRoomForLegend}{Whether to leave room for the color bar/legend}

\item{forceColorsInRange}{Whether or not to force colors in the plotted
range. Useful if you have a value outside of the range or that is NA after
being transformed via the scale that you still want to plot at the edge of
the color scale}

\item{pch}{see \code{\link{graphics::points}}}

\item{colorName}{What plotting parameter controls the color of the points.}

\item{...}{Arguments to the plot function. For example, if \code{pch==19}, set to
'bg', but if \code{pch==1}, set to 'col'.
\code{\link{graphics::plot.default}}}

\item{scaleFunInverse}{Inverse of \code{scaleFun}. For example, exp}
}
\description{
Column major ordering, row and col start at index 1
}
\details{
A wrapper around fields::image.plot for easy plotting of points with color
and easy use of nonlinear color scales in base R.
}
\examples{
set.seed(123)
x = matrix(rnorm(100), ncol=2)
z = rnorm(50)
splot(x, z, colScale=redBlueDivCols, centerScale=TRUE)
}
\seealso{
\code{\link{fields::image.plot}},
\code{\link{graphics::plot.default}}
}
\author{
John Paige
}
