% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splott.R
\name{squilt}
\alias{squilt}
\title{Scaled quilt plots}
\usage{
squilt(
  x,
  y,
  z = NULL,
  zlim = NULL,
  col = NULL,
  colScale = blueGreenYellowSeqCols,
  nlevel = 64,
  scaleFun = I,
  scaleFunInverse = I,
  center = 0,
  centerScale = FALSE,
  compressed = FALSE,
  n.ticks = 5,
  min.n = 5,
  ticks = NULL,
  tickLabels = NULL,
  forceColorsInRange = FALSE,
  otherAxis.args = NULL,
  ...
)
}
\arguments{
\item{x}{Horizontal spatial coordinates, or 2d matrix with x-y coords}

\item{y}{Vertical spatial coordinates, or response is x is 2d matrix}

\item{z}{Response (prior to scaling), or null if x is 2d matrix. Determines
color of the points}

\item{zlim}{range of the response (prior to scaling)}

\item{col}{a vector of colors determining the color scale for plotting
(generally it is recommended to use the \code{colScale} argument instead)}

\item{colScale}{a function for generating a color scale. See
\code{?colorScales}}

\item{nlevel}{Number of colors to include in generated color scale if \code{col}
not specified}

\item{scaleFun}{How to scale the color scale. For example, log}

\item{scaleFunInverse}{Inverse of \code{scaleFun}. For example, exp}

\item{center}{Value at which to center the color scale if \code{centerScale} set
to TRUE}

\item{centerScale}{Whether to center the color scale at \code{center} (e.g. for
diverging color scales)}

\item{compressed}{Rather than cutting off a color scale to make the center at
the desired point, compress the scale on the shorter side so it changes color
faster.}

\item{n.ticks}{Approximate number of ticks in color scale. See
\code{\link{pretty}}}

\item{min.n}{Minimum number of ticks in color scale. See
\code{\link{pretty}}}

\item{ticks}{Unscaled colorscale tick marks}

\item{tickLabels}{labels for ticks}

\item{forceColorsInRange}{Whether or not to force colors in the plotted
range. Useful if you have a value outside of the range or that is NA after
being transformed via the scale that you still want to plot at the edge of
the color scale}

\item{otherAxis.args}{Arguments to pass to \code{fields::image.plot} via
\code{axis.args} (aside from \code{at} and \code{labels})}

\item{...}{Arguments to the \code{fields::image.plot} and \code{fields::quilt.plot}
functions.}

\item{new}{Whether to make a new plot or add to the current one}
}
\value{
Output from \code{\link{fields::quilt.plot}}
}
\description{
Wrapper around fields::quilt.plot for spatial quilt plots with easy automatic
color scaling
}
\details{
A wrapper around fields::quilt.plot for easy nonlinear color scales.
}
\examples{
set.seed(123)
x = matrix(rnorm(100), ncol=2)
z = rnorm(50)
splot(x, z, colScale=redBlueDivCols, centerScale=TRUE)
}
\seealso{
\code{\link{fields::image.plot}}, \code{\link{fields::quilt.plot}},
\code{\link{colorScales}}
}
\author{
John Paige
}
